Generate config.h using Meson
The configuration file is an important header which was automatically
generated by autotools. Now that we moved to Meson, all the job has to
be done with it. For this reason, include/meson.build file has been
added and buildsystem has been adapted in order to have the same result
we had with autotools.

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>